CR4MMST tests for Tully and Rabi
Note that these tests only check that the code runs -- the results are
pegged to what actually came out of the tests the first time. So the results
could (hypothetically) be wrong, although they look reasonable. There's a
note above the tests to this effect to remind me of this fact should numbers
come out weird.

A full test of these models based on previously reported results will be the
check that the code is actually correct.